# app/Http/Requests/GeneralDescendantRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Contracts\Validation\Validator;

class GeneralDescendantRequest extends FormRequest
{
    public $validator = null;

    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, mixed>
     */
    public function rules()
    {
        return [
            'name' => 'required|string',
            'gender' => 'required|boolean',
            'parent_id' => 'required|integer'
        ];
    }
    public function failedValidation(Validator $validator)
    {
        $this->validator = $validator;
    }
}